Updated Route + Added Slug + Updated Filter + Minor Bug fixed...

+++ ProcessManagement/Controllers/GroupController.cs
@@ -10,11 +10,10 @@ using ProcessManagement.Services;
 using ProcessManagement.Filters;
 namespace ProcessManagement.Controllers
 {
-    
+    [LanguageFilter]
     public class GroupController : BaseController
     {
         ///=============================================================================================
-        PMSEntities db = new PMSEntities();
         GroupService groupService = new GroupService();
         CommonService commonService = new CommonService();
         ParticipateService participateService = new ParticipateService();
@@ -52,17 +51,18 @@ namespace ProcessManagement.Controllers
             participateService.createParticipate(idUser, group.Id, part, true);
 
             //set flash
-            SetFlash("Success", "Created Group Successfully");
-            return RedirectToAction("Index", "Group");
+            SetFlash(FlashType.Success, "Created Group Successfully");
+            return RedirectToRoute("GroupLocalizedDefault", new { controller = "group", action = "index"});
         }
+
+        [Authorize]
         [GroupAuthorize]
-        public ActionResult Show(int id)
+        public ActionResult Show(string userslug, string groupslug)
         {
             //Tìm group theo id
-            Group group = groupService.findGroup(id);
+            Group group = groupService.findGroup(userslug , groupslug);
             if (group == null) return HttpNotFound();
 
-            Session["idGroup"] = group.Id;
             //Tìm tất cả member thuộc group đó
             var ListParticipant = participateService.findMembersInGroup(group.Id);
             ViewData["ListParticipant"] = ListParticipant;
@@ -70,81 +70,76 @@ namespace ProcessManagement.Controllers
             ViewData["ListProcess"] = processService.findProcess(group.Id);
             return View(group);
         }
-        //public ActionResult AddMember(int? id)
-        //{
-        //    string userid = User.Identity.GetUserId();
-        //    Group group = db.Groups.Find(id);
-        //    if (group == null) return HttpNotFound();
 
-        //    var ListParticipant = db.Participates.Where(x => x.IdGroup == id).ToList();
-        //    ViewData["ListParticipant"] = ListParticipant;
-        //    List<string> userInGroup = new List<string>();
-        //    foreach (var item in ListParticipant)
-        //    {
-        //        userInGroup.Add(item.IdUser);
-        //    }
-
-        //    //string temp = String.Join(", ", userInGroup); 
-        //    ViewData["ListUser"] = db.AspNetUsers.Where(x => !userInGroup.Contains(x.Id)).OrderByDescending(x => x.Id).ToList();
-        //    Session["idGroup"] = id;
-        //    return View(group);
-        //}
         [Authorize]
-        [HttpPost]
-        public ActionResult AddMember(Group model, List<string> adduser)
-        {
-            Group group = groupService.findGroup(model.Id);
-            if (group == null) return HttpNotFound();
-
-            //Add members
-            participateService.addMembers(model, adduser);
-            TempData["UserSetting"] = "ABC";
-            SetFlash("Success", "Added Members Successfully");
-            return RedirectToAction("settings", new { id = model.Id });
-        }
-        [Authorize]
+        [GroupAuthorize]
         //[RoleAuthorize("Admin")]
-        public ActionResult Settings(int id)
+        public ActionResult Setting(string userslug, string groupslug)
         {
             string idUser = User.Identity.GetUserId();
-            Group group = groupService.findGroup(id);
+            Group group = groupService.findGroup(userslug,groupslug);
             if (group == null) return HttpNotFound();
             //Tìm tất cả member thuộc group đó
             var ListParticipant = participateService.findMembersInGroup(group.Id);
             ViewData["ListParticipant"] = ListParticipant;
-
+            ViewData["Group"] = group;
             //Tìm tất cả member không thuộc group đó
             ViewData["ListUser"] = participateService.findMembersNotInGroup(ListParticipant);
             ViewData["Roles"] = participateService.getRoleOfMember(idUser, group.Id);
             return View(group);
         }
+
         [Authorize]
+        [GroupAuthorize]
+        [HttpPost]
+        public ActionResult AddMember(List<string> adduser)
+        {
+            int groupid = (int)Session["groupid"];
+            Group group = groupService.findGroup(groupid);
+            if (group == null) return HttpNotFound();
+
+            //Add members
+            participateService.addMembers(group, adduser);
+            SetTab(TabType.UserSetting);
+            SetFlash(FlashType.Success, "Added Members Successfully");
+            return RedirectToRoute("GroupControlLocalizedDefault", new { action = "setting", userslug = group.ownerSlug, groupslug = group.groupSlug });
+        }
+
+        [Authorize]
+        [GroupAuthorize]
         [HttpPost]
         public ActionResult Edit(Group model, HttpPostedFileBase ImageGr)
         {
-            Group group = groupService.findGroup(model.Id);
+            int groupid = (int)Session["groupid"];
+            Group group = groupService.findGroup(groupid);
             if (group == null) return HttpNotFound();
-
-            commonService.saveAvatarGroup(model,ImageGr, Server.MapPath("~/Content/images/workspace/"));
-            groupService.editGroup(model);
-            TempData["GaneralSetting"] = "ABC";
-            return RedirectToAction("Settings", new { id = model.Id });
+            
+            //edit
+            Group groupEdit = groupService.compareBeforeEdit(group, model);
+            commonService.saveAvatarGroup(groupEdit, ImageGr, Server.MapPath("~/Content/images/workspace/"));
+            groupService.editGroup(groupEdit);
+            SetTab(TabType.GeneralSetting);
+            return RedirectToRoute("GroupControlLocalizedDefault", new { action = "setting", userslug = group.ownerSlug, groupslug = group.groupSlug });
         }
         [Authorize]
+        [GroupAuthorize]
         [HttpGet]
-        public ActionResult RemoveAvatar(Group model)
+        public ActionResult RemoveAvatar()
         {
-            Group group = groupService.findGroup(model.Id);
+            int groupid = (int)Session["groupid"];
+            Group group = groupService.findGroup(groupid);
             if (group == null) return HttpNotFound();
-            groupService.removeAvatar(model);
-            TempData["GaneralSetting"] = "ABC";
-            SetFlash("Success", "Removed Avatar Successfully");
-            return RedirectToAction("Settings", new { id = model.Id });
+            groupService.removeAvatar(group);
+            SetTab(TabType.GeneralSetting);
+            SetFlash(FlashType.Success, "Removed Avatar Successfully");
+            return RedirectToRoute("GroupControlLocalizedDefault", new { action = "setting", userslug = group.ownerSlug, groupslug = group.groupSlug });
         }
         [Authorize]
-        public ActionResult Remove(int id)
+        [GroupAuthorize]
+        public ActionResult Remove()
         {
-            Group group = groupService.findGroup(id);
+            int groupid = (int)Session["groupid"];
+            Group group = groupService.findGroup(groupid);
             var groupName = group.Name;
             ////////////////////////////////////////////////////////////////////////////
             //TODO
@@ -154,56 +149,113 @@ namespace ProcessManagement.Controllers
             
             //remove group (bao gồm remove participant,process,step)
             groupService.removeGroup(group);
-            SetFlash("Success", "Removed Group "+ groupName + " Successfully");
+            SetFlash(FlashType.Success, "Removed Group "+ groupName + " Successfully");
             return RedirectToAction("Index");
         }
         
         [Authorize]
+        [GroupAuthorize]
         public ActionResult DeleteMember(Participate model)
         {
             Participate user = participateService.findMemberInGroup(model.Id);
             if (user == null) return HttpNotFound();
             var userName = user.AspNetUser.UserName;
-            var groupId = user.IdGroup;
-            participateService.removeUserInGroup(user);
-            TempData["UserSetting"] = "ABC";
-            SetFlash("Success", "Removed "+ userName + " Successfully");
-            return RedirectToAction("Settings", new { id = groupId });
+            int groupId = (int)Session["groupid"];
+            Group group = groupService.findGroup(groupId);
+            ////////////////////////////////////////////////////////////////////////////
+            //NOTE
+            //Không được xóa đi owner,không được xóa thành viên khách group
+            ///////////////////////////////////////////////////////////////////////////
+            if (user.IsOwner)
+                SetFlash(FlashType.Fail, "You cant remove owner");
+            else
+            {
+                participateService.removeUserInGroup(user);
+                SetFlash(FlashType.Success, "Removed " + userName + " Successfully");
+            }
+
+            SetTab(TabType.UserSetting);
+            return RedirectToRoute("GroupControlLocalizedDefault", new { action = "setting", userslug = group.ownerSlug, groupslug = group.groupSlug });
         }
         [Authorize]
+        [GroupAuthorize]
         public ActionResult EditRoleMember(int id)
         {
- 
+            
             Participate user = participateService.findMemberInGroup(id);
             if (user == null) return HttpNotFound();
+
+            int groupid = (int)Session["groupid"];
+            Group group = groupService.findGroup(groupid);
+            if (group == null) return HttpNotFound();
+           
+
+
+            ////////////////////////////////////////////////////////////////////////////
+            //NOTE
+            //Tại owner không được sửa quyền sửa role của chính nó
+            //TODO
+            //Hỏi customer lại về vụ nè
+            //Nếu sửa role của chính nó thì sẽ báo lỗi gì??tạm thời return 404
+            ///////////////////////////////////////////////////////////////////////////
+            if (user.IsOwner)
+            {
+                SetFlash(FlashType.Fail, "Owner cant change their role");
+                SetTab(TabType.UserSetting);
+                return RedirectToRoute("GroupControlLocalizedDefault", new { action = "setting", userslug = group.ownerSlug, groupslug = group.groupSlug });
+            }
             return View(user);
         }
+
         [Authorize]
+        [GroupAuthorize]
         [HttpPost]
         public ActionResult EditRoleMember(Participate model)
         {
+            var groupId = (int)Session["groupid"];
+            Group group = groupService.findGroup(groupId);
             Participate user = participateService.findMemberInGroup(model.Id);
             if (user == null) return HttpNotFound();
+            
             //chỉnh sửa role của 1 user
-            participateService.editRoleUser(model);
 
-            var groupId = user.IdGroup;
-            TempData["UserSetting"] = "ABC";
-            SetFlash("Success", "Edited Role of "+user.AspNetUser.UserName+" Successfully");
-            return RedirectToAction("Settings", new { id = groupId });
+            if (!user.IsOwner)
+            {
+                participateService.editRoleUser(model);
+                SetFlash(FlashType.Success, "Edited Role of "+user.AspNetUser.UserName+" Successfully");
+            }
+            else
+                SetFlash(FlashType.Fail, "Owner cant change their role");
+
+            SetTab(TabType.UserSetting);
+            return RedirectToRoute("GroupControlLocalizedDefault", new { action = "setting", userslug = group.ownerSlug, groupslug = group.groupSlug });
         }
         [Authorize]
         public ActionResult MemberLeaveGroup(Participate model)
         {
+            string idUser = User.Identity.GetUserId();
             Participate user = participateService.findMemberInGroup(model.Id);
             if (user == null) return HttpNotFound();
             var groupName = user.Group.Name;
 
             //xóa thành viên trong group
-            participateService.removeUserInGroup(user);
+            ////////////////////////////////////////////////////////////////////////////
+            //NOTE
+            //Chỉ xóa được chính bản thân mình ra khỏi group
+            ///////////////////////////////////////////////////////////////////////////
+            if (user.IdUser == idUser)
+            {
+                participateService.removeUserInGroup(user);
+                SetFlash(FlashType.Success, "Left Group " + groupName + " Successfully");
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                SetFlash(FlashType.Fail, "Left Group " + groupName + " Failed");
+                SetTab(TabType.AdvancedSetting);
+                return RedirectToRoute("GroupControlLocalizedDefault", new { action = "setting", userslug = user.Group.ownerSlug, groupslug = user.Group.groupSlug });
+            }
             
-            SetFlash("Success", "Left Group " + groupName + " Successfully");
-            return RedirectToAction("Index");
         }
     }
 }
